+++ txcdr-client/app/(auth)/dashboard/AdminSection.tsx
@@ -1,15 +1,23 @@
-import { Pressable, StyleSheet, View } from "react-native";
+import {
+  ActivityIndicator,
+  Pressable,
+  ScrollView,
+  StyleSheet,
+  View,
+} from "react-native";
 import { DText } from "../../../components/styled-rn/DText";
 import { useState } from "react";
 import { ModeButton } from "./ModeButton";
 import { Blue } from "../../../utils/styles/colors";
 import { router } from "expo-router";
 import { ms } from "react-native-size-matters";
+import { SectionProps } from ".";
+import { EventCard } from "./EventCard";
 
-export function AdminPage() {
+export function AdminSection({ events }: SectionProps) {
   const [mode, setMode] = useState<"ACTIVE" | "PAST">("ACTIVE");
   return (
-    <View style={styles.container}>
+    <>
       <View style={styles.modeSelector}>
         <ModeButton
           label="Active"
@@ -22,17 +30,40 @@ export function AdminPage() {
           onPress={() => setMode("PAST")}
         />
       </View>
-      <DText>Event Dashboard</DText>
-      <View style={{ flex: 1 }}>
-        <DText>Filler</DText>
-      </View>
+      {events ? (
+        <ScrollView
+          style={styles.scroller}
+          contentContainerStyle={styles.eventList}
+        >
+          {events.map(
+            (
+              e,
+              i, // TODO: filter by active column
+            ) => (
+              <EventCard
+                id={e.id}
+                registered={e.approved}
+                severity={e.severity}
+                title={e.title}
+                key={e.id}
+              />
+            ),
+          )}
+        </ScrollView>
+      ) : (
+        <ActivityIndicator
+          size="large"
+          style={{
+            justifyContent: "center",
+            height: "100%",
+          }}
+        />
+      )}
       <View style={styles.buttonContainer}>
         <Pressable
           onPress={() => router.push("/dashboard/new-event")}
           style={{
             alignContent: "flex-end",
-            marginRight: ms(20),
-            marginBottom: ms(20),
             backgroundColor: Blue[500],
             borderRadius: 20,
           }}
@@ -48,7 +79,7 @@ export function AdminPage() {
           </DText>
         </Pressable>
       </View>
-    </View>
+    </>
   );
 }
 
@@ -63,9 +94,20 @@ const styles = StyleSheet.create({
     flexDirection: "row",
     gap: ms(10),
     justifyContent: "center",
+    marginTop: ms(20),
   },
   buttonContainer: {
     flexDirection: "row",
     justifyContent: "flex-end",
+  },
+  scroller: {
+    paddingBottom: 50,
+    marginTop: ms(20),
+  },
+  eventList: {
+    alignItems: "center",
+    alignSelf: "center",
+    justifyContent: "center",
+    width: "100%",
   },
 });
